锁后台ota功能查询条件、超级管理员权限、灰度测试功能等

+++ src/main/java/socket/gsm/admin/controller/BinFileController.java
@@ -88,6 +88,10 @@ public class BinFileController extends BaseController{
 		if(id == null){
 			return responseParamFail("参数不合法");
 		}
+		String value = (String) getRequest().getSession().getAttribute("user");
+		if(!LoginUser.checkAdminUser(value)){
+			return responseFail("没有权限");
+		}
         try {
         	int res = binFileService.delete(id);
         	if(res > 0){
@@ -134,7 +138,7 @@ public class BinFileController extends BaseController{
 	@ResponseBody
 	@RequestMapping(value="/editWhileListMac",method=RequestMethod.POST)
     public Object editWhileListMac(Integer id,String whiteListMac){
-		if(id == null || StringUtils.isBlank(whiteListMac)){
+		if(id == null){
 			return responseParamFail("参数不合法");
 		}
 		String value = (String) getRequest().getSession().getAttribute("user");
